chore: renamed old arm control package
