+++ arm_control_old/img/ElbowPositions.png
