+++ arm_control_old/model/meshes/EOAT.stl
