+++ arm_control_old/model/meshes/Link0.stl
