+++ arm_control_old/model/meshes/Link1.stl
