+++ arm_control_old/model/meshes/Link2.stl
