+++ arm_control_old/model/meshes/Link3.stl
