+++ arm_control_old/model/meshes/collision_Link0.stl
