+++ arm_control_old/model/meshes/collision_Link1.stl
