+++ arm_control_old/model/meshes/collision_Link2.stl
