+++ arm_control_old/model/meshes/collision_Link3.stl
