+++ arm_control_old/model/meshes/collision_Link4.stl
