+++ arm_control_old/model/meshes/finger.stl
