+++ arm_control_old/rviz/arm_viz_double.rviz
